covid19.web/requirements.txt
asgiref==3.2.5
Django==3.0.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
pydash==4.7.6
pytz==2019.3
sqlparse==0.3.1
